Create script and regenerate bindings with `bindgen`
There used to be a stale "automatically generated by rust-bindgen"
comment confusing potential contributors how to generate new bindings
when AOSP inevitably adds new functions and structures to this library.

This is very simply fixed by checking in a new script that invokes
`bindgen` with the right parameters such that we maintain some MSRV
compatibility and filter the output on relevant types (and their
dependencies) exclusively.